refactor: DTO 반환하도록 변경
